Succesfully modified the Pack Configuration MSB.
The code can now modify the MSB of the Pack Configuration register. The problem rigth now is reading the old value and the new value of the register (2 bytes) the reading of the LSB byte can be done correctly, but when reading de MSB the result is not good, it's like if i was reading the LSB two times.