Reading data flash succesfully
Now it's possible to read the data flash and display it. The only problemes is that it does not display the two bytes at the same time, it displays the MSB first and after 8 seconds it displays the LSB, and so on. The next step is to fix the modifyDataFlash function, so that it can modify 2 bytes at the same time.