Reading and modification of the flash can be done
The reading and modification of the data flash can be done, but it has 3 problems. First, the reading can read only 1 byte at the time. Second, the modification works only for 1 byte and the biggest problem right now is that the function for modifying the flash is not general, so i am going to be modifing that so it can work with any register of the data flash.